Updated model tests from master to new location and format
